moved diva_experiment process
